requirements.txt
newspaper3k==0.2.8
warcio==1.7.4